Refactor critic model
